backend/requirements.txt
flask
flask-cors
pymongo
python-dotenv
bcrypt
requests
certifi
twilio
google-genai
